Create an Ansible role to add a Windows machine as a Jenkins slave
As a preset for this to work on a system a node has to be created
manually on the target Jenkins and the name and secret have to be set
for each host. Afterwards the script will create a service user and
workspace folder. It will place the slave jar into that folder and register
a service to start the node connection automatically.